// pages/types.ts
export type Role = 
  | 'admin' 
  | 'system_supervisor' 
  | 'planning_officer' 
  | 'head_general' 
  | 'head_raw' 
  | 'head_finished' 
  | 'supervisor_general' 
  | 'supervisor_raw' 
  | 'supervisor_finished' 
  | 'storekeeper_general' 
  | 'storekeeper_raw' 
  | 'storekeeper_finished' 
  | 'cashier' 
  | string;